ADuCM3029. ADC Temperature Polling Mode Example. Interrupts files, serial transmission through the UART is only required for this example

+++ ADuCM3029/Examples/ADC_Temperature_Polling_Mode/inc/interrupts.h
@@ -37,7 +37,7 @@ extern void UART_Int_Handler ( void );
 
 /**@brief Variables.
  */
-extern volatile uint8_t  myState;
+extern volatile uint32_t  myState;
 extern volatile uint8_t  *myPtr;
